fixes some scenarios of value type usage
mainly scenarios in which the address of a value type
is needed. There are still one known issue that need to
 be fixed: when multiple invocations happens on 
the return of methods (returning a value type) C#
compiler reuses the synthetic local variable and
our current code just introduces new local variables
for each of such invocation.

Code that would trigger that are commented out in the test