sort type declarations by dependency before processing
this improves the chances of dependencies of a type to be processed
before the type itself. For instance, in the following code:

class A : B {}
class B {}

'B' will be processed before 'A' leading to code with a more natural
flow, i.e, code that emits the type 'B' followed by code that emits
the type 'A', which is easier to reason about.

Prior to this change, Cecilifier would start processing 'A' and
would find that it needs to process 'B' (at least creating a
'TypeDefinition' representing 'B') and then continue processing 'A'
which would lead to code related to 'A' being mixed with code
related to 'B'.