Move call to GitHub query APIs to retrieve issues/releases to backend
We do that to be able to do authenticated API calls to have a higher rate limit (unauthenticated API calls have a very low rate limit) without
needing to transmit the github token to the client. In the future we may experiment with generating a random key at client, encrypting it with
server public key (from certificate) and sending this key to the server which would use it to encript the github token and sent it back to
client. This way, at least the token would never be transmited in plain text.