June 2023 (#241)
* fixes exception thrown in break statements placed outside switch statements (#232)

* renames local variable to follow convention

* fixes staging server URL

* use 'foreach' like syntax instead of plain 'for' in js.

* Move call to GitHub query APIs to retrieve issues/releases to backend

We do that to be able to do authenticated API calls to have a higher rate limit (unauthenticated API calls have a very low rate limit) without
needing to transmit the github token to the client. In the future we may experiment with generating a random key at client, encrypting it with
server public key (from certificate) and sending this key to the server which would use it to encript the github token and sent it back to
client. This way, at least the token would never be transmited in plain text.

* bump cecilifier version to 2.02

* reverts package update directly in .csproj


instead, updates version in common packages.

* add support for tests to do custom instruction validation during assembly comparison

this can be used when the expected / generated assemblies have minor differences in instruction operands, for example, if local variables names
differs

* change array initialization to more closely match C# compiler optimizations (#222)

basically in somce scenarios C# compiler generates a struct type to hold raw data used to intialize arrays without
setting each array element (bulk copy)

* code cleanup: better parameter name + left over comment removal

* change stackalloc initialization to more closely match C# compiler optimizations (#222)

* removes outdated duplicated test

Other stackalloc tests covers the same code paths as the removed one

---------

Co-authored-by: Adriano Carlos Verona <adriano@unity3d.com>