improves handling of value type instantiation
In various scenarios the code was not handling value type instantiation correctly. This commit fixes a bunch of such
scenarios.

There are still at least one known scenario (passing the resulting of a ObjectCreationExpression for a value type
as an in parameter) which is being mishandled. This will be fixed in a separate commit