improves snippet / cecilified code mapping
more specifically ensures that clicking on the method name on an invocation expression will select the generated
code for the invocation instead of the code that simply loads the target of the invocation.

Also ensures that clicking on arguments on invocations will select the generated code related for the evaluation
of the arguments being passed.

Signed-off-by: Adriano Carlos Verona <adriano.verona@gmail.com>