use test case parameterized values instead of hardcoded ones (#256)
Test was using [2, 1] hardcoded collection expression instead of the parameterized values whence collection
expressions larger than 2 were not being tested.

Since this length is used to determine how the initialization is performed optimized initialization (length >= 3)
was not being tested.

Fixing this revealed a bug in the logic that selects optimized/unoptimized initialization. AFAICS optimization is only
applicable when the target of the collection expression (i.e, a variable assignment/initialization or argument passing)
is a container (array/collection/etc) of primitive types but the code was not taking this into consideration.