reenabled more generic type parameter boxing tests (#263)
there is one failure due to differences in the generated assembly.
One of the diffeerneces is that we do not emit [.readonly](https://github.com/stakx/ecma-335/blob/master/docs/iii.2.3-readonly.md) prefix when
emitting 'ldelema' instruction for arrays of generic type paramters.

A second error has to do with validation of fields of `PrivateImplementationDetails` emitted types.
Cecilifier is generating the name of the static field by calling PrivateImplementationDetailsGenerator.GetOrCreateInitializationBackingFieldVariableName()
passing an expression with an array of values used to initialize the array, something like: "new int[] {1, 2, 3}" and this is converted to a byte array
and is hashed inside GetOrCreateInitializationBackingFieldVariableName() while roslyn takes the actual int[] and convert it to byte (not the C# string
representation), i.e, something like: { 0x01, 0x00, 0x00, 0x00, 0x02, 0x00, 0x00, 0x00, 0x03,0x00, 0x00, 0x00 }"