improves association of fields/methods with its owning type (#354)
The list of methods/fields of a type depends on the type the list of methods/fields of the next method.

If the type defines methods/fields, the token of the first method/field is used to define the list of methods/field.
All methods/fields with a token >= the first method/field token and <= first method/field token of the next type
are associated with the type.

Unfortunately neither the "Common Language Infrastructure" docs nor the Sytem.Reflection.Metadata documentation
is clear wrt that (at least IMHO).

My understanding on how methods/fields gets associated with a type is derived from dumping metadata tables
for varios scenarios and observing the actual values (a simple, but very effective way to dump the metadata
tables is to use the tool from git@github.com:kant2002/MetadataDumper.git)