refactor: improve forwarded method declararion handling (#360)
Unfortunatelly the code mapping from IMethodSymbol -> MethodAttributes enum is incomplete and
cecilifier code just assumes that the attributes are incorrect and resets them (at least
with the correct value). At some point we should either improve the mapping and remove
this reset code or remove the mapping (i.e. rely on the reseting code) and make it
clear in the generated code that this reseting is happening